+++ go/apps/host/host.go
@@ -32,7 +32,8 @@ import (
 	"github.com/jlmucb/cloudproxy/go/tao"
 	"github.com/jlmucb/cloudproxy/go/util"
 	"github.com/jlmucb/cloudproxy/go/util/options"
-	"github.com/golang/crypto/ssh/terminal"
+	// "github.com/golang/crypto/ssh/terminal"
+	"golang.org/x/crypto/ssh/terminal"
 )
 
 var opts = []options.Option{
