+++ go/apps/tao_admin/tao_admin.go
@@ -33,7 +33,8 @@ import (
 	"github.com/jlmucb/cloudproxy/go/tao"
 	"github.com/jlmucb/cloudproxy/go/tao/auth"
 	"github.com/jlmucb/cloudproxy/go/util/options"
-	"github.com/golang/crypto/ssh/terminal"
+	// "github.com/golang/crypto/ssh/terminal"
+	"golang.org/x/crypto/ssh/terminal"
 )
 
 var opts = []options.Option{
